Separated the RamExecutor header file
