+++ src/AstTuner.h
@@ -18,7 +18,6 @@
 
 #include "AstTransformer.h"
 #include "Global.h"
-#include "RamExecutor.h"
 
 namespace souffle {
