+++ src/RamExecutor.h
@@ -20,81 +20,21 @@
 #include "RamProgram.h"
 #include "RamRelation.h"
 #include "SymbolTable.h"
-#include "Util.h"
 
 #include "RamStatement.h"
 
-#include <functional>
-//#include <map>
 #include <ostream>
-#include <string>
 #include <vector>
 
 namespace souffle {
 
-/** forward declaration */
-class RamStatement;
-class RamInsert;
-
-class EvalContext {
-    std::vector<const RamDomain*> data;
-    std::vector<RamDomain>* returnValues = nullptr;
-    std::vector<bool>* returnErrors = nullptr;
-    const std::vector<RamDomain>* args = nullptr;
-
-public:
-    EvalContext(size_t size = 0) : data(size) {}
-
-    const RamDomain*& operator[](size_t index) {
-        return data[index];
-    }
-
-    const RamDomain* const& operator[](size_t index) const {
-        return data[index];
-    }
-
-    std::vector<RamDomain>& getReturnValues() const {
-        return *returnValues;
-    }
-
-    void setReturnValues(std::vector<RamDomain>& retVals) {
-        returnValues = &retVals;
-    }
-
-    void addReturnValue(RamDomain val, bool err = false) {
-        assert(returnValues != nullptr && returnErrors != nullptr);
-        returnValues->push_back(val);
-        returnErrors->push_back(err);
-    }
-
-    std::vector<bool>& getReturnErrors() const {
-        return *returnErrors;
-    }
-
-    void setReturnErrors(std::vector<bool>& retErrs) {
-        returnErrors = &retErrs;
-    }
-
-    const std::vector<RamDomain>& getArguments() const {
-        return *args;
-    }
-
-    void setArguments(const std::vector<RamDomain>& a) {
-        args = &a;
-    }
-
-    RamDomain getArgument(size_t i) const {
-        assert(args != nullptr && i < args->size() && "argument out of range");
-        return (*args)[i];
-    }
-};
-
 /**
- * An abstract base class for entities capable of processing a RAM program.
+ * An abstract base class for executing/synthesizing RAM programs
  */
+
 class RamExecutor {
 protected:
-    /** An optional stream to print logging information to */
+    /** An optional stream to print logging information to an output stream */
     std::ostream* report;
 
 public:
@@ -143,254 +83,11 @@ public:
      */
     virtual void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
             const std::vector<RamDomain>& arguments, std::vector<RamDomain>& returnValues,
-            std::vector<bool>& returnErrors) = 0; 
-
+            std::vector<bool>& returnErrors) const = 0; 
     /**
      * Runs the given statement on the given environment.
      */
     virtual void applyOn(const RamProgram& prog, RamEnvironment& env, RamData* data) const = 0;
-};
-
-/**
- * A class representing the order of elements.
- */
-class Order {
-    /** The covered order */
-    std::vector<unsigned> order;
-
-public:
-    static Order getIdentity(unsigned size) {
-        Order res;
-        for (unsigned i = 0; i < size; i++) {
-            res.append(i);
-        }
-        return res;
-    }
-
-    void append(unsigned pos) {
-        order.push_back(pos);
-    }
-
-    unsigned operator[](unsigned index) const {
-        return order[index];
-    }
-
-    std::size_t size() const {
-        return order.size();
-    }
-
-    bool isComplete() const {
-        for (size_t i = 0; i < order.size(); i++) {
-            if (!contains(order, i)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    const std::vector<unsigned>& getOrder() const {
-        return order;
-    }
-
-    void print(std::ostream& out) const {
-        out << order;
-    }
-
-    friend std::ostream& operator<<(std::ostream& out, const Order& order) {
-        order.print(out);
-        return out;
-    }
-};
-
-/**
- * The summary to be returned from a statement executor.
- */
-struct ExecutionSummary {
-    Order order;
-    long time;
-};
-
-/** Defines the type of execution strategies */
-typedef std::function<ExecutionSummary(const RamInsert&, RamEnvironment& env, std::ostream*)>
-        QueryExecutionStrategy;
-
-// -- example strategies --
-
-/** With this strategy queries will be processed as they are stated by the user */
-extern const QueryExecutionStrategy DirectExecution;
-
-/** With this strategy queries will be dynamically rescheduled before each execution */
-extern const QueryExecutionStrategy ScheduledExecution;
-
-/**
- * An interpreter based implementation of a RAM executor. The RAM program will
- * be processed within the callers process. Before every query operation, an
- * optional scheduling step will be conducted.
- */
-class RamGuidedInterpreter : public RamExecutor {
-    /** The executor processing a query */
-    QueryExecutionStrategy queryStrategy;
-
-public:
-    /** A constructor accepting a query executor strategy */
-    RamGuidedInterpreter(const QueryExecutionStrategy& queryStrategy = ScheduledExecution)
-            : queryStrategy(queryStrategy) {}
-
-    /**
-     * The implementation of the interpreter applying the given program
-     * on the given environment.
-     */
-    void applyOn(const RamProgram& prog, RamEnvironment& env, RamData* data) const override;
-
-    /**
-     * Runs a subroutine of a RamProgram
-     */
-    virtual void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
-            const std::vector<RamDomain>& arguments, std::vector<RamDomain>& returnValues,
-            std::vector<bool>& returnErrors) override;
-};
-
-/**
- * An interpreter based implementation of a RAM executor. The RAM program will
- * be processed within the callers process. In this version, no scheduling
- * will be conducted.
- */
-struct RamInterpreter : public RamGuidedInterpreter {
-    /** A constructor setting the query policy for the base class */
-    RamInterpreter() : RamGuidedInterpreter(DirectExecution){};
-};
-
-/**
- * A RAM executor based on the creation and compilation of an executable conducting
- * the actual computation.
- */
-class RamCompiler : public RamExecutor {
-private:
-    std::string compileCmd;
-
-public:
-    /** A simple constructor */
-    RamCompiler(const std::string& compileCmd) : compileCmd(compileCmd) {}
-
-    /**
-     * Generates the code for the given ram statement.The target file
-     * name is either set by the corresponding member field or will
-     * be determined randomly. The chosen file-name will be returned.
-     */
-    std::string generateCode(const SymbolTable& symTable, const RamProgram& prog,
-            const std::string& filename = "", const int index = -1) const;
-
-    /**
-     * Generates the code for the given ram statement.The target file
-     * name is either set by the corresponding member field or will
-     * be determined randomly. The chosen file-name will be returned.
-     */
-    std::string compileToLibrary(const SymbolTable& symTable, const RamProgram& prog,
-            const std::string& filename = "default", const int index = -1) const;
-
-    /**
-     * Compiles the given statement to a binary file. The target file
-     * name is either set by the corresponding member field or will
-     * be determined randomly. The chosen file-name will be returned.
-     * Note that this uses the generateCode method for code generation.
-     */
-    std::string compileToBinary(const SymbolTable& symTable, const RamProgram& prog,
-            const std::string& filename = "", const int index = -1) const;
-
-    /**
-     * Compiles the given statement to a binary file. The target file
-     * name is either set by the corresponding member field or will
-     * be determined randomly. The environment after execution will be returned.
-     * Note that this uses the compileToBinary method for code compilation.
-     */
-    std::string executeBinary(const SymbolTable& symTable, const RamProgram& prog,
-            const std::string& filename = "", const int index = -1) const;
-
-    /**
-     * The actual implementation of this executor encoding the given
-     * program into a source file, compiling and executing it.
-     */
-    void applyOn(const RamProgram& prog, RamEnvironment& env, RamData* data) const override;
-
-    /**
-     * Compiles a subroutine into a separate method in the compiled
-     * program
-     */
-    // void compileSubroutine(std::string name, const RamStatement& stmt, IndexMap& indices, std::ostream& os)
-    // const;
-    virtual void executeSubroutine(RamEnvironment& env, const RamStatement& stmt,
-            const std::vector<RamDomain>& arguments, std::vector<RamDomain>& returnValues,
-            std::vector<bool>& returnErrors) override {
-        // nothing to do here
-    }
-};
-
-/**
- * A singleton which provides a mapping from strings to unique valid CPP identifiers.
- */
-class CPPIdentifierMap {
-public:
-    /**
-     * Obtains the singleton instance.
-     */
-    static CPPIdentifierMap& getInstance() {
-        if (instance == nullptr) {
-            instance = new CPPIdentifierMap();
-        }
-        return *instance;
-    }
-
-    /**
-     * Given a string, returns its corresponding unique valid identifier;
-     */
-    static std::string getIdentifier(const std::string& name) {
-        return getInstance().identifier(name);
-    }
-
-    ~CPPIdentifierMap() = default;
-
-private:
-    CPPIdentifierMap() {}
-
-    static CPPIdentifierMap* instance;
-
-    /**
-     * Instance method for getIdentifier above.
-     */
-    const std::string identifier(const std::string& name) {
-        auto it = identifiers.find(name);
-        if (it != identifiers.end()) {
-            return it->second;
-        }
-        // strip leading numbers
-        unsigned int i;
-        for (i = 0; i < name.length(); ++i) {
-            if (isalnum(name.at(i)) || name.at(i) == '_') {
-                break;
-            }
-        }
-        std::string id;
-        for (auto ch : std::to_string(identifiers.size() + 1) + '_' + name.substr(i)) {
-            // alphanumeric characters are allowed
-            if (isalnum(ch)) {
-                id += ch;
-            }
-            // all other characters are replaced by an underscore, except when
-            // the previous character was an underscore as double underscores
-            // in identifiers are reserved by the standard
-            else if (id.size() == 0 || id.back() != '_') {
-                id += '_';
-            }
-        }
-        // most compilers have a limit of 2048 characters (if they have a limit at all) for
-        // identifiers; we use half of that for safety
-        id = id.substr(0, 1024);
-        identifiers.insert(std::make_pair(name, id));
-        return id;
-    }
-
-    // The map of identifiers.
-    std::map<const std::string, const std::string> identifiers;
 };
 
 }  // end of namespace souffle
