+++ src/RamSynthesiser.cpp
@@ -14,7 +14,7 @@
  *
  ***********************************************************************/
 
-#include "RamExecutor.h"
+#include "RamSynthesiser.h"
 #include "AstRelation.h"
 #include "AstVisitor.h"
 #include "BinaryConstraintOps.h"
@@ -46,6 +46,74 @@
 #endif
 
 namespace souffle {
+
+/**
+ * A singleton which provides a mapping from strings to unique valid CPP identifiers.
+ */
+class CPPIdentifierMap {
+public:
+    /**
+     * Obtains the singleton instance.
+     */
+    static CPPIdentifierMap& getInstance() {
+        if (instance == nullptr) {
+            instance = new CPPIdentifierMap();
+        }
+        return *instance;
+    }
+
+    /**
+     * Given a string, returns its corresponding unique valid identifier;
+     */
+    static std::string getIdentifier(const std::string& name) {
+        return getInstance().identifier(name);
+    }
+
+    ~CPPIdentifierMap() = default;
+
+private:
+    CPPIdentifierMap() {}
+
+    static CPPIdentifierMap* instance;
+
+    /**
+     * Instance method for getIdentifier above.
+     */
+    const std::string identifier(const std::string& name) {
+        auto it = identifiers.find(name);
+        if (it != identifiers.end()) {
+            return it->second;
+        }
+        // strip leading numbers
+        unsigned int i;
+        for (i = 0; i < name.length(); ++i) {
+            if (isalnum(name.at(i)) || name.at(i) == '_') {
+                break;
+            }
+        }
+        std::string id;
+        for (auto ch : std::to_string(identifiers.size() + 1) + '_' + name.substr(i)) {
+            // alphanumeric characters are allowed
+            if (isalnum(ch)) {
+                id += ch;
+            }
+            // all other characters are replaced by an underscore, except when
+            // the previous character was an underscore as double underscores
+            // in identifiers are reserved by the standard
+            else if (id.size() == 0 || id.back() != '_') {
+                id += '_';
+            }
+        }
+        // most compilers have a limit of 2048 characters (if they have a limit at all) for
+        // identifiers; we use half of that for safety
+        id = id.substr(0, 1024);
+        identifiers.insert(std::make_pair(name, id));
+        return id;
+    }
+
+    // The map of identifiers.
+    std::map<const std::string, const std::string> identifiers;
+};
 
 // See the CPPIdentifierMap, (it is a singleton class).
 CPPIdentifierMap* CPPIdentifierMap::instance = nullptr;
